+++ testsuite/tests/driver/T4437.hs
@@ -39,6 +39,7 @@ expectedGhcOnlyExtensions :: [String]
 expectedGhcOnlyExtensions = ["RelaxedLayout",
                              "AlternativeLayoutRule",
                              "AlternativeLayoutRuleTransitional",
+                             "DerivingVia",
                              "EmptyDataDeriving",
                              "GeneralisedNewtypeDeriving",
                              "QuantifiedConstraints"]